This is the first Implementation of a web scraping algorithm.
The algorithm first checks the site passed and if it satisfies the check, a url is set up with the information(make, model) and a network request is sent to the url.

The wanted information is then extracted and sent to a text file, it is further extracted and send to another file.

The data is then sent through a cleaning method to remove unwanted divs.